Started looking into python-docx and moved backend code to helper functions accordingly.
