+++ prov-api/routes/explore/explore.py
@@ -10,7 +10,8 @@ from typing import Dict, Any
 from config import Config, get_settings
 
 
-from ProvenaInterfaces.RegistryAPI import ItemSubType, ItemBase, SeededItem
+from ProvenaInterfaces.RegistryAPI import ItemSubType, ItemBase 
+from helpers.generate_report_helpers import validate_node_id, generate_report_helper
 
 router = APIRouter()
 
@@ -321,21 +322,6 @@ async def effected_agents(
     )
 
 
-async def _validate_node_id(node_id: str, item_subtype: ItemSubType, request_style: RequestStyle, config: Config) -> None:
-
-    if item_subtype == ItemSubType.MODEL_RUN:
-        response = await validate_model_run_id(id=node_id, request_style=request_style, config=config)
-    elif item_subtype == ItemSubType.STUDY:
-        response = await validate_study_id(id=node_id, request_style=request_style, config=config)
-    else:
-        raise HTTPException(status_code=400, detail="Unsupported item subtype for validation.")
-
-    if isinstance(response, str):
-        raise HTTPException(status_code=400, detail=f"Validation error with provided {item_subtype.value}: {response}")
-    
-    if isinstance(response, SeededItem):
-        raise HTTPException(status_code=400, detail="Seeded item cannot be used for this query!")
-
 @router.get("/generate/report", response_model = SimpleDummyResponse, operation_id="export_prov_graph")
 # Setup the endpoint inputs.
 async def export_graph(
@@ -344,12 +330,8 @@ async def export_graph(
     item_subtype: ItemSubType,
     roles: ProtectedRole = Depends(read_user_protected_role_dependency),
     config: Config = Depends(get_settings)
-) -> SimpleDummyResponse: 
-    
-    from helpers.prov_helpers import generate_report_helper 
-
-    assert config.depth_upper_limit, "config no depth upper limit."
-    
+) -> SimpleDummyResponse:
+         
     # Create the request style, here we assert where the HTTP request came from. 
     request_style: RequestStyle = RequestStyle(
         user_direct=roles.user, service_account=None
@@ -360,7 +342,7 @@ async def export_graph(
         raise HTTPException(status_code=400, detail="Invalid input depth provided. Input depth is only from 1 to 3")
     
     # do checks accordingly. 
-    await _validate_node_id(node_id=node_id, item_subtype=item_subtype, request_style=request_style, config=config)
+    await validate_node_id(node_id=node_id, item_subtype=item_subtype, request_style=request_style, config=config)
 
     # This acting as a shared dictionary across both edge cases and in the helper function in the edge cases.
     all_filtered_responses: Dict[str, List[ItemBase]] = {"inputs": [], "outputs": [], "model_runs": []}
@@ -382,14 +364,15 @@ async def export_graph(
         # This may contain more than one model run.
         upstream_model_run_response = await explore_downstream(starting_id=node_id, depth=1, config=config)
         assert upstream_model_run_response.graph, "Node collections not found!"
-        model_run_nodes: List[Any] = upstream_model_run_response.graph.get('nodes', [])
+
+        model_run_nodes: List[ItemBase] = upstream_model_run_response.graph.get('nodes', [])
 
         # Now branch out with the model runs in here and explore them as above (depth 1-3 upstream and depth 1 downstream)
         # The model run will be the new updated starting id.
 
         for model_run_node in model_run_nodes:
-            if model_run_node.get('item_subtype') == ItemSubType.MODEL_RUN:
-                model_run_id = model_run_node.get('id')
+            if model_run_node.item_subtype == ItemSubType.MODEL_RUN:
+                model_run_id = model_run_node.id
                 # Extract the upstream and downstream entities from this node. 
                 await generate_report_helper(starting_id=model_run_id, upstream_depth=depth, shared_responses=all_filtered_responses, config=config)
